requirements.txt
BloomFilter==0.1.0
watchdog==3.0.0
yara-python==4.3.1
